StockOrdering backend:
- updated stock orders table : made price a float type instead of integer
- updated type script interface for back-end, created type: stockOrderDetails

+++ build5-TypescriptBuild/servers/stock_ordering/src/database/database.sql
@@ -6,7 +6,7 @@ Create TABLE stock_orders (
     order_type TEXT NOT NULL,
     ticker TEXT NOT NULL,
     quantity INTEGER NOT NULL,
-    price INTEGER NOT NULL,
+    price float NOT NULL,
     order_time TIMESTAMP NOT NULL,
     order_status TEXT NOT NULL
 )
